+++ webhook.go
@@ -7,15 +7,17 @@ import (
 
 // webhook リクエスト
 type WebhookRequest struct {
-	Key    string `json:"key"`
-	RoomId string `json:"room_id"`
+	Key    *string `json:"key,omitempty"`
+	RoomId string  `json:"room_id"`
 }
 
 // webhook レスポンス
 type WebhookResponse struct {
-	Allowed    *bool  `json:"allowed"`
-	WebhookUrl string `json:"auth_webhook_url"`
-	Reason     string `json:"reason"`
+	Allowed       bool          `json:"allowed"`
+	IceServers    []interface{} `json:"iceServers,omitempty"`
+	WebhookUrl    *string       `json:"auth_webhook_url,omitempty"`
+	Reason        string        `json:"reason,omitempty"`
+	AuthzMetadata interface{}   `json:"authz_metadata"`
 }
 
 // TODO(kdxu): 送信するデータを吟味する
@@ -24,12 +26,7 @@ type TwoAuthnRequest struct {
 	AuthnMetadata interface{} `json:"authn_metadata"`
 }
 
-type TwoAuthnResponse struct {
-	Allowed       *bool       `json:"allowed"`
-	AuthzMetadata interface{} `json:"authz_metadata"`
-}
-
-func AuthWebhookRequest(key string, roomId string, metadata interface{}, host string) (interface{}, error) {
+func AuthWebhookRequest(key *string, roomId string, metadata interface{}, host string) (*WebhookResponse, error) {
 	webhookReq := &WebhookRequest{Key: key, RoomId: roomId}
 	respBytes, err := PostRequest(Options.AuthWebhookUrl, webhookReq)
 	whResp := WebhookResponse{}
@@ -37,24 +34,24 @@ func AuthWebhookRequest(key string, roomId string, metadata interface{}, host st
 	if err != nil {
 		return nil, err
 	}
-	if !*whResp.Allowed {
+	if !whResp.Allowed {
 		logger.Info("authn webhook not allowed, resp=", &whResp)
 		return nil, errors.New("Not Allowed")
 	}
-	if whResp.WebhookUrl != "" {
-		respBytes, err := PostRequest(whResp.WebhookUrl, &TwoAuthnRequest{Host: &host, AuthnMetadata: metadata})
-		twoAuthnResp := TwoAuthnResponse{}
+	if whResp.WebhookUrl != nil {
+		respBytes, err := PostRequest(*whResp.WebhookUrl, &TwoAuthnRequest{Host: &host, AuthnMetadata: metadata})
+		twoAuthnResp := WebhookResponse{IceServers: whResp.IceServers}
 		err = json.Unmarshal(respBytes, &twoAuthnResp)
 		if err != nil {
 			return nil, err
 		}
-		if !*twoAuthnResp.Allowed {
+		if !twoAuthnResp.Allowed {
 			logger.Info("two authn webhook not allowed, resp=", &twoAuthnResp)
 			return nil, errors.New("Not Allowed")
 		}
 		logger.Info("two authn webhook allowed, resp=", &twoAuthnResp)
-		return twoAuthnResp.AuthzMetadata, nil
+		return &twoAuthnResp, nil
 	}
 	logger.Info("auth webhook allowed, resp=", whResp)
-	return nil, nil
+	return &whResp, nil
 }